task: add anchor IDs for PlanRel and ReferenceRel
these IDs are a way to identify a PlanRel (anchor target) and reference
it from a ReferenceRel (anchor source), without relying on something as
brittle as position in a list. The `subtree ordinal` field is left in to
allow it to be used to more quickly identify a `PlanRel`. I am also
using it to distinguish a `ReferenceRel` that points to a `PlanRel` in
another plan (when subtree_ordinal is 0) as opposed to in the same plan
(subtree_ordinal is greater than 0).