task: improved `DecomposeStats`
Added new steps: to capture `PARSE` when stringifying, and to
disambiguate between executing a pushback that *only* has a
`SkyResultRel` (MATERIALIZE) from executing some work delegated in the
pushback (TRANSLATE).

Also, added fields for the active engine and for start/stop timestamps
to better capture which steps are contained in others.