task: accommodated feedback from PR
This accommodates 2 primary changes:
  1. change the name of the reference and anchor fields
  2. simplify the nesting of the old and new fields into a `oneof` to a
     simple addition of the anchor field

The first change is to be more descriptive: `planrel_reference` makes it
clear what the anchor points to, `reference_anchor` can be a reusable
field name that specifies it is an anchor designed to be used by a
`ReferenceRel` ("referencerel_anchor" seems unnecessarily annoying to
read).

The second change is to make migration more natural by deprecating the
existing field and adding the new field in a way that is backwards
compatible.

PR: 726